+++ src/git/git.ts
@@ -23,9 +23,39 @@ export * from './remotes/provider';
 let git: IGit;
 
 const defaultBlameParams = ['blame', '--root', '--incremental'];
+
 // Using %x00 codes because some shells seem to try to expand things if not
-const defaultLogParams = ['log', '--name-status', '--full-history', '-M', '--format=%x3c%x2ff%x3e%n%x3cr%x3e %H%n%x3ca%x3e %an%n%x3ce%x3e %ae%n%x3cd%x3e %at%n%x3cp%x3e %P%n%x3cs%x3e%n%B%x3c%x2fs%x3e%n%x3cf%x3e'];
-const defaultStashParams = ['stash', 'list', '--name-status', '--full-history', '-M', '--format=%x3c%x2ff%x3e%n%x3cr%x3e %H%n%x3cd%x3e %at%n%x3cl%x3e %gd%n%x3cs%x3e%n%B%x3c%x2fs%x3e%n%x3cf%x3e'];
+const lb = '%x3c'; // `%x${'<'.charCodeAt(0).toString(16)}`;
+const rb = '%x3e'; // `%x${'>'.charCodeAt(0).toString(16)}`;
+const sl = '%x2f'; // `%x${'/'.charCodeAt(0).toString(16)}`;
+
+const logFormat = [
+    `${lb}${sl}f${rb}`,
+    `${lb}r${rb} %H`,   // ref
+    `${lb}a${rb} %an`,  // author
+    `${lb}e${rb} %ae`,  // email
+    `${lb}d${rb} %at`,  // date
+    `${lb}p${rb} %P`,   // parents
+    `${lb}s${rb}`,
+    `%B`,               // summary
+    `${lb}${sl}s${rb}`,
+    `${lb}f${rb}`
+].join('%n');
+
+const defaultLogParams = ['log', '--name-status', '--full-history', '-M', `--format=${logFormat}`];
+
+const stashFormat = [
+    `${lb}${sl}f${rb}`,
+    `${lb}r${rb} %H`,   // ref
+    `${lb}d${rb} %at`,  // date
+    `${lb}l${rb} %gd`,  // reflog-selector
+    `${lb}s${rb}`,
+    `%B`,               // summary
+    `${lb}${sl}s${rb}`,
+    `${lb}f${rb}`
+].join('%n');
+
+const defaultStashParams = ['stash', 'list', '--name-status', '--full-history', '-M', `--format=${stashFormat}`];
 
 const GitWarnings = [
     /Not a git repository/,