+++ src/git/parsers/logParser.ts
@@ -106,6 +106,11 @@ export class GitLogParser {
                             entry.summary += `\n${line}`;
                         }
                     }
+
+                    if (entry.summary !== undefined) {
+                        // Remove the trailing newline
+                        entry.summary = entry.summary.slice(0, -1);
+                    }
                     break;
 
                 case 102: // 'f': // files
@@ -223,7 +228,7 @@ export class GitLogParser {
                 entry.author!,
                 entry.email,
                 new Date(entry.date! as any * 1000),
-                entry.summary!,
+                entry.summary === undefined ? '' : entry.summary,
                 relativeFileName,
                 entry.fileStatuses || [],
                 entry.status,